# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AppRestaurant",
        "description": "Listado de URI´S de la API restaurante",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/dishes": {
            "get": {
                "tags": [
                    "Dish"
                ],
                "summary": "Listado de los dishes",
                "operationId": "ddf841d9ba35d917bd22a504c13b2f72",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aderson"
                                                    },
                                                    "imagen_path": {
                                                        "type": "string",
                                                        "example": "1.png"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pasta"
                                                    },
                                                    "price": {
                                                        "type": "numeric",
                                                        "example": "0"
                                                    },
                                                    "restaurant_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dish"
                ],
                "summary": "Registrar la información del dish",
                "operationId": "3fc0252a7de74dac3747d26739c844cb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "imagen_path": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "numeric"
                                            },
                                            "restaurant_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Aderson Felix",
                                    "imagen_path": "P1.png",
                                    "description": "datos",
                                    "price": "10",
                                    "restaurant_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "imagen_path": {
                                            "type": "string",
                                            "example": "1.png"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "pasta"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "restaurant_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The description field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/dishes/{id}": {
            "get": {
                "tags": [
                    "Dish"
                ],
                "summary": "Mostrar la información de un dish",
                "operationId": "07b6fe7c8dc79b35e66b58956b97b893",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "imagen_path": {
                                            "type": "string",
                                            "example": "1.png"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "pasta"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "restaurant_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dish"
                ],
                "summary": "Eliminar la información de un dish",
                "operationId": "57bb2309cba8e67241d5d84d8ceb6beb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/dishes/{id}": {
            "put": {
                "tags": [
                    "Dish"
                ],
                "summary": "Actualizar la información del dish",
                "operationId": "3fe03e4e62306720eb569b0a830cce40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "imagen_path": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "numeric"
                                            },
                                            "restaurant_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Aderson Felix",
                                    "imagen_path": "P1.png",
                                    "description": "datos",
                                    "price": "10",
                                    "restaurant_id": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "imagen_path": {
                                            "type": "string",
                                            "example": "1.png"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "pasta"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "restaurant_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The description field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}